Use settings in jdbc url for createdb and dropdb
`SequelRails::Storage::Postgres` assumes that username, host, port and
password is set in the config. Otherwise it connects to localhost
regardless of the contents of the jdbc URL which breaks e.i. `rake
db:test:prepare`. This change parses the contents of jdbc URL (for
postgresql) and sets these fields if they do not have any value.